+++ packages/kit/src/core/config/index.spec.js
@@ -2,11 +2,59 @@ import { join } from 'path';
 import { fileURLToPath } from 'url';
 import { test } from 'uvu';
 import * as assert from 'uvu/assert';
-import { remove_keys } from '../../utils/object.js';
 import { validate_config, load_config } from './index.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = join(__filename, '..');
+
+/**
+ * mutates and remove keys from an object when check callback returns true
+ * @param {Record<string, any>} o any object
+ * @param {([key, value]: [string, any]) => boolean} check callback with access
+ * 		to the key-value pair and returns a boolean that decides the deletion of key
+ */
+function remove_keys(o, check) {
+	for (const key in o) {
+		if (!Object.hasOwnProperty.call(o, key)) continue;
+		if (check([key, o[key]])) delete o[key];
+		const nested = typeof o[key] === 'object' && !Array.isArray(o[key]);
+		if (nested) remove_keys(o[key], check);
+	}
+}
+
+const directive_defaults = {
+	'child-src': undefined,
+	'default-src': undefined,
+	'frame-src': undefined,
+	'worker-src': undefined,
+	'connect-src': undefined,
+	'font-src': undefined,
+	'img-src': undefined,
+	'manifest-src': undefined,
+	'media-src': undefined,
+	'object-src': undefined,
+	'prefetch-src': undefined,
+	'script-src': undefined,
+	'script-src-elem': undefined,
+	'script-src-attr': undefined,
+	'style-src': undefined,
+	'style-src-elem': undefined,
+	'style-src-attr': undefined,
+	'base-uri': undefined,
+	sandbox: undefined,
+	'form-action': undefined,
+	'frame-ancestors': undefined,
+	'navigate-to': undefined,
+	'report-uri': undefined,
+	'report-to': undefined,
+	'require-trusted-types-for': undefined,
+	'trusted-types': undefined,
+	'upgrade-insecure-requests': false,
+	'require-sri-for': undefined,
+	'block-all-mixed-content': false,
+	'plugin-types': undefined,
+	referrer: undefined
+};
 
 const get_defaults = (prefix = '') => ({
 	extensions: ['.svelte'],
@@ -21,41 +69,13 @@ const get_defaults = (prefix = '') => ({
 		},
 		csp: {
 			mode: 'auto',
-			directives: {
-				'child-src': undefined,
-				'default-src': undefined,
-				'frame-src': undefined,
-				'worker-src': undefined,
-				'connect-src': undefined,
-				'font-src': undefined,
-				'img-src': undefined,
-				'manifest-src': undefined,
-				'media-src': undefined,
-				'object-src': undefined,
-				'prefetch-src': undefined,
-				'script-src': undefined,
-				'script-src-elem': undefined,
-				'script-src-attr': undefined,
-				'style-src': undefined,
-				'style-src-elem': undefined,
-				'style-src-attr': undefined,
-				'base-uri': undefined,
-				sandbox: undefined,
-				'form-action': undefined,
-				'frame-ancestors': undefined,
-				'navigate-to': undefined,
-				'report-uri': undefined,
-				'report-to': undefined,
-				'require-trusted-types-for': undefined,
-				'trusted-types': undefined,
-				'upgrade-insecure-requests': false,
-				'require-sri-for': undefined,
-				'block-all-mixed-content': false,
-				'plugin-types': undefined,
-				referrer: undefined
-			}
+			directives: directive_defaults,
+			reportOnly: directive_defaults
 		},
-		endpointExtensions: ['.js', '.ts'],
+		endpointExtensions: undefined,
+		env: {
+			publicPrefix: 'PUBLIC_'
+		},
 		files: {
 			assets: join(prefix, 'static'),
 			hooks: join(prefix, 'src/hooks'),
@@ -65,7 +85,6 @@ const get_defaults = (prefix = '') => ({
 			serviceWorker: join(prefix, 'src/service-worker'),
 			template: join(prefix, 'src/app.html')
 		},
-		floc: false,
 		headers: undefined,
 		host: undefined,
 		hydrate: undefined,
@@ -74,6 +93,7 @@ const get_defaults = (prefix = '') => ({
 			parameter: '_method',
 			allowed: []
 		},
+		moduleExtensions: ['.js', '.ts'],
 		outDir: join(prefix, '.svelte-kit'),
 		package: {
 			dir: 'package',
@@ -95,6 +115,7 @@ const get_defaults = (prefix = '') => ({
 			entries: ['*'],
 			force: undefined,
 			onError: 'fail',
+			origin: 'http://sveltekit-prerender',
 			pages: undefined
 		},
 		protocol: undefined,
@@ -105,7 +126,9 @@ const get_defaults = (prefix = '') => ({
 		version: {
 			name: Date.now().toString(),
 			pollInterval: 0
-		}
+		},
+		// TODO cleanup for 1.0
+		vite: undefined
 	}
 });
 
@@ -115,7 +138,6 @@ test('fills in defaults', () => {
 	assert.equal(validated.kit.package.exports(''), true);
 	assert.equal(validated.kit.package.files(''), true);
 	assert.equal(validated.kit.serviceWorker.files(''), true);
-	assert.equal(validated.kit.vite(), {});
 
 	remove_keys(validated, ([, v]) => typeof v === 'function');
 
@@ -181,7 +203,6 @@ test('fills in partial blanks', () => {
 	assert.equal(validated.kit.package.exports(''), true);
 	assert.equal(validated.kit.package.files(''), true);
 	assert.equal(validated.kit.serviceWorker.files(''), true);
-	assert.equal(validated.kit.vite(), {});
 
 	remove_keys(validated, ([, v]) => typeof v === 'function');