+++ packages/kit/test/apps/basics/src/routes/errors/page-endpoint/get-explicit.js
@@ -1,0 +1,6 @@
+import { FancyError } from './_shared.js';
+
+export const GET = () => ({
+	status: 400,
+	body: new FancyError('oops')
+});