+++ packages/kit/test/apps/basics/src/routes/src/[...anything].js
@@ -1,0 +1,6 @@
+/** @type {import('./__types/[...anything]').RequestHandler} */
+export function GET() {
+	return {
+		body: 'dynamically rendered file'
+	};
+}