+++ packages/kit/test/prerendering/basics/svelte.config.js
@@ -1,4 +1,3 @@
-import path from 'path';
 import adapter from '../../../../adapter-static/index.js';
 
 /** @type {import('@sveltejs/kit').Config} */
@@ -8,19 +7,8 @@ const config = {
 
 		prerender: {
 			default: true,
-			onError: 'continue'
-		},
-
-		vite: {
-			build: {
-				minify: false
-			},
-			clearScreen: false,
-			server: {
-				fs: {
-					allow: [path.resolve('../../../src')]
-				}
-			}
+			onError: 'continue',
+			origin: 'http://example.com'
 		}
 	}
 };