+++ sites/kit.svelte.dev/src/routes/content.json.js
@@ -18,7 +18,7 @@ const categories = [
 	}
 ];
 
-export function get() {
+export function GET() {
 	/** @type {import('$lib/search/types').Block[]} */
 	const blocks = [];